add dtheme.html, mtheme.html
also minor fix

+++ upgrade/upgrade.php
@@ -2025,10 +2025,18 @@ $mysqlMainDb = ' . quote($mysqlMainDb) . ';
 
 
         // drop stale full text indexes
-        Database::get()->query("ALTER TABLE document DROP INDEX document");        
-        Database::get()->query("ALTER TABLE course_units DROP INDEX course_units_title");
-        Database::get()->query("ALTER TABLE course_units DROP INDEX course_units_comments");
-        Database::get()->query("ALTER TABLE unit_resources DROP INDEX unit_resources_title");
+        if (DBHelper::indexExists('document', 'document')) {
+            Database::get()->query("ALTER TABLE document DROP INDEX document");
+        }
+        if (DBHelper::indexExists('course_units', 'course_units_title')) {
+            Database::get()->query("ALTER TABLE course_units DROP INDEX course_units_title");
+        }
+        if (DBHelper::indexExists('course_units', 'course_units_comments')) {
+            Database::get()->query("ALTER TABLE course_units DROP INDEX course_units_comments");    
+        }
+        if (DBHelper::indexExists('unit_resources', 'unit_resources_title')) {
+            Database::get()->query("ALTER TABLE unit_resources DROP INDEX unit_resources_title");
+        }
         
         // // ----------------------------------
         // creation of indexes
